feat(registry): extract submission risk helpers into submission-risk-lib
Move draft and direct-content risk scoring into submission-risk-lib with consolidated Vitest coverage.

Preserve the public @heyclaude/registry/submission-risk import surface.
